Title: n개의 숫자 중 최대 2개, Time: 6ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 240213/n개의 숫자 중 최대 2개/two-max-of-n-num.cpp
@@ -2,7 +2,7 @@
 using namespace std;
 
 int main() {
-    int arr[100] = {0};
+    long long int arr[100] = {0};
     int n;
 
     cin >> n;
@@ -23,7 +23,6 @@ int main() {
                 arr[j+1] = temp;
             }            
         }
-
     }
 
     cout << arr[0] << " " << arr[1];
